Content amends for 966, expand on health conditions / application process
Hypothesis for 966 content amends:

We've seen that people are confused by the application process. We think this is because it's not clear on the GOV.UK start page that there is an eligibility checker, then the main application form, then (in BAU cases) a referral for a work capability assessment (WCA) where the claimant is sent a ESA50 questionnaire for more information on their condition. We've also seen from the feedback form that users are anxious to provide more information on their condition, and we think this is because they don't know they'll have the opportunity to elaborate later in the ESA50.

So if we add more information on the application process at key points in the journey (eligibility start page, eligibility results page, conditions page)

We'll see that claimants have a clearer understanding of the application process and will not be anxious to provide more information in the application form. We'll know this is true through user testing with claimants and third party organisations.